UPBGE: Fix for compiling under Ubuntu 16.04
Avoid non defined operator "*=" for Mathfu. It expects Math vector and
float

+++ source/gameengine/Ketsji/KX_ObstacleSimulation.cpp
@@ -421,7 +421,7 @@ void KX_ObstacleSimulationTOI::AdjustObstacleVelocity(KX_Obstacle *activeObst, K
 	mt::vec2 dv = activeObst->nvel - activeObst->vel;
 	const float ds = dv.Length();
 	if (ds > maxDeltaSpeed || ds < -maxDeltaSpeed) {
-		dv *= fabs(maxDeltaSpeed / ds);
+		dv *= fabsf(maxDeltaSpeed / ds);
 	}
 	const mt::vec2 vel = activeObst->vel + dv;
